Optional parameter to allow manual definition of zip type.
This allows users of the open function to manually define the zip type (which is useful for when files do not have the correct extension). This allows zips that have been uploaded by php (which depending on the platform will have no extension or a .tmp extension) to be extracted directly.